# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Homeschool Hive API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface AuthUser {
  id: string;
  email?: string | null;
  firstName?: string | null;
  lastName?: string | null;
  lastNameInitial?: string | null;
  displayName?: string | null;
  profileImageUrl?: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface UpdateProfileRequest {
  firstName: string;
  lastNameInitial: string;
}

export interface Tag {
  id: number;
  name: string;
  slug: string;
  color: string;
}

export interface Reaction {
  emoji: string;
  count: number;
  userReacted: boolean;
}

export interface ReactionsResponse {
  reactions: Reaction[];
}

export type LinkPriceRange =
  | "free"
  | "under_10"
  | "under_50"
  | "under_100"
  | "under_500"
  | "over_500";

export type LinkStatus = "pending" | "approved" | "rejected";

export interface Link {
  id: number;
  url: string;
  title: string;
  summary?: string | null;
  comment?: string | null;
  thumbnailUrl?: string | null;
  createdAt: string;
  updatedAt: string;
  userId: string;
  status?: LinkStatus | null;
  priceRange?: LinkPriceRange | null;
}

export type LinkWithDetails = Link & {
  tags: Tag[];
  reactions: Reaction[];
  user: AuthUser;
  totalReactions: number;
};

export interface CreateLinkRequest {
  url: string;
  title: string;
  summary?: string | null;
  comment?: string | null;
  tagIds?: number[];
  priceRange?: LinkPriceRange | null;
}

export interface UpdateLinkRequest {
  url?: string;
  title?: string;
  summary?: string | null;
  comment?: string | null;
  tagIds?: number[];
  priceRange?: LinkPriceRange | null;
}

export interface AddReactionRequest {
  emoji: string;
}

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListLinksParams = {
  search?: string;
  tags?: string;
  sortBy?: ListLinksSortBy;
  priceRange?: LinkPriceRange;
};

export type ListLinksSortBy =
  (typeof ListLinksSortBy)[keyof typeof ListLinksSortBy];

export const ListLinksSortBy = {
  newest: "newest",
  oldest: "oldest",
  most_reactions: "most_reactions",
  title: "title",
} as const;